+++ nba2k_shotchart_project/ref_hackbase/HackBase/renderer.h
@@ -10,7 +10,7 @@ public:
 	virtual void EndScene() = 0;
 	virtual void DrawRect(int x, int y, int w, int h, Color color) = 0;
 	virtual void DrawBorder(int x, int y, int w, int h, int d, Color color) = 0;
-	virtual void DrawLine(int x1, int y1, int x2, int y2, float size, bool antialias, Color color) = 0;
+	virtual void DrawLine(float x1, float y1, float x2, float y2, int size, bool antialias, Color color) = 0;
 	virtual void DrawText(int x, int y, Color FontColor, char *Text, ...) = 0;
 	virtual void DrawCircle(int x, int y, float radius, int width, UINT samples, Color color) = 0;
 	virtual int GetWidth() = 0;
